chore - update typos and gradle to version 6.7

+++ net.tangly.ports/src/main/java/net/tangly/crm/ports/CrmTsvHdl.java
@@ -198,7 +198,7 @@ public class CrmTsvHdl {
     }
 
     static void addActivities(Interaction entity, Provider<Activity> activities) {
-        entity.addAll(activities.items().stream().filter(o -> ((Long) ReflectionUtilities.get(o, "interactionFoid")).longValue() == entity.oid())
+        entity.addAll(activities.items().stream().filter(o -> (Long) ReflectionUtilities.get(o, "interactionFoid") == entity.oid())
                 .collect(Collectors.toList()));
     }
